+++ test_app/objects/agg_many1.c
@@ -10,7 +10,7 @@
 
 using namespace std;
 
-int calculate_agg_many1( obj_loc_t obj_loc, int id )
+int calculate_agg_many1( obj_loc_t obj_loc, long id )
 {
     for ( int i = 0 ; i < NUM_MANY1 ; i++ ) 
     {
