+++ test_app/objects/numeric.c
@@ -10,7 +10,7 @@
 
 using namespace std;
 
-int calculate_numeric( obj_loc_t obj_loc, int id )
+int calculate_numeric( obj_loc_t obj_loc, long id )
 {
     // Does nothing.
